migrate webmock to test-only environment
