+++ Gemfile
@@ -69,6 +69,9 @@ group :test, :development do
   gem 'pry-rails'
   gem 'coveralls', require: false
   gem 'poltergeist'
+end
+
+group :test do
   gem 'webmock'
 end
